Added Code to Handle the Command Line Arguments
Command Line Arguments can now be handled and will check for the three options (-f, -t, and -s).

The contents following each are now stored in a variable.

The filePath is retrieved from -f and then is used to create the various Shape3D objects depending on the first string for each line.